requirements.txt
Django==1.10.5
django-debug-toolbar==1.6
django-filter==1.0.1
django-webpack-loader==0.4.1
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
Markdown==2.6.7
mongoengine==0.11.0
pkg-resources==0.0.0
PyJWT==1.4.2
pymongo==3.4.0
six==1.10.0
sqlparse==0.2.2
